OPTIMIZATION: Optimized bsp_sensor_getAngle().

+++ CyclePOV/CyclePOV/tasks/inc/bikeInfo.h
@@ -29,7 +29,7 @@ typedef struct{
 	float stripeLength;
 	uint16_t numLeds;
 
-	float period;
+	uint32_t period;
 	float velocity;
 
 	Stripes stripes;
